relocated bottomNav component from test folder into Buyer profile folder
