+++ client/src/layouts/RootLayout.jsx
@@ -1,7 +1,7 @@
 import Navbar from '../pages/Navbar/Navbar';
 import { Outlet } from 'react-router-dom';
 import { useAuthContext } from '../hooks/useAuthContext';
-import BottomNav from '../pages/Profile/_tests_/BottomNav';
+import BottomNav from '../pages/Profile/Buyer/BottomNav';
 
 
 export default function RootLayout() {
